--- WebMarketDomain/Entities/Interfaces/IEntity.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace WebMarketDomain.Entities.Interfaces
-{
-    public interface IEntity
-    {
-        int Id { get; set; }
-
-
-    }
-}
