+++ WebMarketDomain/Entityes/Interfaces/IEntity.cs
@@ -1,0 +1,10 @@
+﻿
+namespace WebMarketDomain.Entityes.Interfaces
+{
+    public interface IEntity
+    {
+        int Id { get; set; }
+
+
+    }
+}
